Move to dyanmic memory for state entries
This fixes #3.

We were using stack memory and expecting it to be around after function
exit. This was magically working forever, but no longer works on more
recent compilers.

This moves the table to being a vector of pointers which we then
allocate and deallocate on each run using new/delete.

While changing this we moved to using `std::sort` on the vector instead
of `qsort`. And improving the sorting of IP addresses as well.

Signed-off-by: Phil Dibowitz <phil@ipom.com>